+++ crates/teloxide-core/src/types/transaction_partner.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 
-use crate::types::{PaidMedia, RevenueWithdrawalState, User};
+use crate::types::{Gift, PaidMedia, RevenueWithdrawalState, Seconds, User};
 
 /// This object describes the source of a transaction, or its recipient for
 /// outgoing transactions.
@@ -11,7 +11,7 @@ use crate::types::{PaidMedia, RevenueWithdrawalState, User};
 #[serde(tag = "type")]
 pub enum TransactionPartner {
     Fragment(TransactionPartnerFragment),
-    User(TransactionPartnerUser),
+    User(Box<TransactionPartnerUser>),
     TelegramAds,
     TelegramApi(TransactionPartnerTelegramApi),
     Other,
@@ -37,11 +37,17 @@ pub struct TransactionPartnerUser {
     /// Bot-specified invoice payload.
     pub invoice_payload: Option<String>,
 
+    /// The duration of the paid subscription.
+    pub subscription_period: Option<Seconds>,
+
     /// Information about the paid media bought by the user.
     pub paid_media: Option<Vec<PaidMedia>>,
 
     /// Bot-specified paid media payload
     pub paid_media_payload: Option<String>,
+
+    /// The gift sent to the user by the bot
+    pub gift: Option<Gift>,
 }
 
 /// Describes a transaction with payment for paid broadcasting.